requirements.txt
asgiref==3.7.2
cffi==1.16.0
Django 
django-cors-headers==4.3.1
django-extensions==3.2.3
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
numpy==1.26.4
openslide-python==1.3.1
pillow==10.2.0
pycparser==2.21
PyJWT==1.7.1
PyMySQL==1.1.0
pytz==2023.3.post1
pyvips
sqlparse
typing_extensions==4.9.0
tzdata==2024.1